+++ lib/src/playable.dart
@@ -1,3 +1,4 @@
+import 'package:assets_audio_player/assets_audio_player.dart';
 import 'package:flutter/foundation.dart';
 
 class Playable {}
@@ -9,20 +10,18 @@ enum AudioType {
   asset,
 }
 
-extension AudioTypeDescription on AudioType {
-  String description() {
-    switch (this) {
-      case AudioType.network:
-        return "network";
-      case AudioType.liveStream:
-        return "liveStream";
-      case AudioType.file:
-        return "file";
-      case AudioType.asset:
-        return "asset";
-    }
-    return null;
+String audioTypeDescription(AudioType audioType) {
+  switch (audioType) {
+    case AudioType.network:
+      return "network";
+    case AudioType.liveStream:
+      return "liveStream";
+    case AudioType.file:
+      return "file";
+    case AudioType.asset:
+      return "asset";
   }
+  return null;
 }
 
 enum ImageType {
@@ -31,18 +30,16 @@ enum ImageType {
   asset,
 }
 
-extension ImageTypeDescription on ImageType {
-  String description() {
-    switch (this) {
-      case ImageType.network:
-        return "network";
-      case ImageType.file:
-        return "file";
-      case ImageType.asset:
-        return "asset";
-    }
-    return null;
+String imageTypeDescription(ImageType imageType) {
+  switch (imageType) {
+    case ImageType.network:
+      return "network";
+    case ImageType.file:
+      return "file";
+    case ImageType.asset:
+      return "asset";
   }
+  return null;
 }
 
 @immutable
@@ -54,10 +51,12 @@ class MetasImage {
   const MetasImage.network(this.path)
       : type = ImageType.network,
         package = null;
+
   const MetasImage.asset(
     this.path, {
     this.package,
   }) : type = ImageType.asset;
+
   const MetasImage.file(this.path)
       : type = ImageType.file,
         package = null;
@@ -104,26 +103,56 @@ class Metas {
   @override
   int get hashCode =>
       title.hashCode ^ artist.hashCode ^ album.hashCode ^ image.hashCode;
+
+  Metas copyWith({
+    String title,
+    String artist,
+    String album,
+    Map<String, dynamic> extra,
+    MetasImage image,
+  }) {
+    return new Metas(
+      title: title ?? this.title,
+      artist: artist ?? this.artist,
+      album: album ?? this.album,
+      extra: extra ?? this.extra,
+      image: image ?? this.image,
+    );
+  }
 }
 
-@immutable
 class Audio implements Playable {
   final String path;
   final String package;
   final AudioType audioType;
-  final Metas metas;
+  Metas _metas;
 
-  const Audio(this.path, {this.metas, this.package})
-      : audioType = AudioType.asset;
-  const Audio.file(this.path, {this.metas})
+  Metas get metas => _metas;
+
+  Audio._({
+    this.path,
+    this.package,
+    this.audioType,
+    Metas metas,
+  }) : _metas = metas;
+
+  Audio(this.path, {Metas metas, this.package})
+      : audioType = AudioType.asset,
+        _metas = metas;
+
+  Audio.file(this.path, {Metas metas})
       : audioType = AudioType.file,
         package = null;
-  const Audio.network(this.path, {this.metas})
+
+  Audio.network(this.path, {Metas metas})
       : audioType = AudioType.network,
-        package = null;
-  const Audio.liveStream(this.path, {this.metas})
+        package = null,
+        _metas = metas;
+
+  Audio.liveStream(this.path, {Metas metas})
       : audioType = AudioType.liveStream,
-        package = null;
+        package = null,
+        _metas = metas;
 
   @override
   bool operator ==(Object other) =>
@@ -138,6 +167,40 @@ class Audio implements Playable {
   @override
   int get hashCode =>
       path.hashCode ^ package.hashCode ^ audioType.hashCode ^ metas.hashCode;
+
+  void updateMetas({
+    AssetsAudioPlayer player,
+    String title,
+    String artist,
+    String album,
+    Map<String, dynamic> extra,
+    MetasImage image,
+  }) {
+    this._metas = (_metas ?? Metas()).copyWith(
+      title: title,
+      artist: artist,
+      album: album,
+      extra: extra,
+      image: image,
+    );
+    if (player != null) {
+      player.onAudioUpdated(this);
+    }
+  }
+
+  Audio copyWith({
+    String path,
+    String package,
+    AudioType audioType,
+    Metas metas,
+  }) {
+    return Audio._(
+      path: path ?? this.path,
+      package: package ?? this.package,
+      audioType: audioType ?? this.audioType,
+      metas: metas ?? this._metas,
+    );
+  }
 }
 
 class Playlist implements Playable {
@@ -179,14 +242,17 @@ class Playlist implements Playable {
   int get hashCode => audios.hashCode ^ startIndex.hashCode;
 }
 
-void writeAudioMetasInto(Map<String, dynamic> params, Metas metas) {
-  if (metas.title != null) params["song.title"] = metas.title;
-  if (metas.artist != null) params["song.artist"] = metas.artist;
-  if (metas.album != null) params["song.album"] = metas.album;
-  if (metas.image != null) {
-    params["song.image"] = metas.image.path;
-    params["song.imageType"] = metas.image.type.description();
-    if (metas.image.package != null)
-      params["song.imagePackage"] = metas.image.package;
+void writeAudioMetasInto(
+    Map<String, dynamic> params, /* nullable */ Metas metas) {
+  if (metas != null) {
+    if (metas.title != null) params["song.title"] = metas.title;
+    if (metas.artist != null) params["song.artist"] = metas.artist;
+    if (metas.album != null) params["song.album"] = metas.album;
+    if (metas.image != null) {
+      params["song.image"] = metas.image.path;
+      params["song.imageType"] = imageTypeDescription(metas.image.type);
+      if (metas.image.package != null)
+        params["song.imagePackage"] = metas.image.package;
+    }
   }
 }